Add offchain proposal data to needs my vote api
- Update getNeedsMyVoteProposals to query for offchain proposals matching onchain proposal IDs
- Create fetchOffchainProposalsMap utility function to reduce code duplication between getProposals and getNeedsMyVoteProposals
- Refactor getProposals to use the new utility function
- Add object cover to voted and not voted list avatars
