+++ CLAUDE.md
@@ -1,13 +1,26 @@
 # Claude Configuration
 
 ## Code Formatting
+
 This project uses Prettier for code formatting. Please format all code changes using the project's Prettier configuration.
 
-Available formatting commands:
+## Code Style
+
+This project uses TypeScript for code style. Please use TypeScript for all code changes.
+Keep the order of imports consistent, with the following order:
+
+1. React and Next.js imports
+2. External imports
+3. Absolute imports
+4. Local imports or relative imports
+
+## Available formatting commands:
+
 - `yarn prettier-src` - Format all files in src/
 - `yarn check-prettier` - Check if files are properly formatted
 
 ## Project Scripts
+
 - `yarn dev` - Start development server
 - `yarn build` - Build for production
 - `yarn lint` - Run ESLint
@@ -15,7 +28,9 @@ Available formatting commands:
 - `yarn test` - Run tests
 
 ## Before committing
+
 Always run these commands before committing:
+
 1. `yarn prettier-src` - Format code
 2. `yarn lint` - Check for linting errors
 3. `yarn typecheck` - Check for TypeScript errors
