+++ src/components/Votes/ProposalVotesList/ProposalSingleVote.tsx
@@ -80,7 +80,7 @@ export function ProposalSingleVote({ vote }: { vote: Vote }) {
                     <img
                       src={vote.voterMetadata.image}
                       alt="avatar"
-                      className="w-full h-full object-contain"
+                      className="w-full h-full object-cover"
                     />
                   </div>
                 ) : (
